requirements.txt
pyTelegramBotAPI==4.14.0
psutil==5.9.6
Pillow
pycaw
pystray>=0.19.0
plyer>=2.1.0
